fix(plinth-submissions): split preview/prod generators
A single Main.hs with `#ifdef PREVIEW` wrote to both 1.45 and 1.61
submission paths. When built under `cabal.project.preview`, the preview
run silently overwrote every 1.45 .uplc with output from the 1.61
plutus-tx-plugin, while each submission's metrics.json kept the older
1.45-plugin numbers — a cross-contaminated state that was easy to miss.

Extract shared `writeCodeToFile` into `Cape.WritePlc`, wire two separate
executables — `plinth-submissions` (1.45 only) and `plinth-submissions-preview`
(1.61 only) — each reachable through its own cabal project. Neither
target can touch the other's paths.

Regenerate the existing 1.45 .uplc artefacts with their honest 1.45-plugin
output, refresh the two_party_escrow 1.61 metrics under the current
evaluator, and add a missing metrics.json for linear_vesting 1.61
(never committed in #152).
