+++ submissions/fibonacci/Plinth_1.45.0.0_Unisay/fibonacci.uplc
@@ -6,19 +6,27 @@
       (lam
         n-1
         (force
-          (case
-            (constr
-              0
-              [ [ (builtin lessThanEqualsInteger) n-1 ] (con integer 1) ]
-              (delay n-1)
+          (force
+            [
+              [
+                [
+                  (force (builtin ifThenElse))
+                  [ [ (builtin lessThanEqualsInteger) n-1 ] (con integer 1) ]
+                ]
+                (delay (delay n-1))
+              ]
               (delay
-                (case
-                  (constr 0 (con integer 0) (con integer 1) (con integer 2) n-1)
-                  fibonacciLoop-0
+                (delay
+                  [
+                    [
+                      [ [ fibonacciLoop-0 (con integer 0) ] (con integer 1) ]
+                      (con integer 2)
+                    ]
+                    n-1
+                  ]
                 )
               )
-            )
-            (force (builtin ifThenElse))
+            ]
           )
         )
       )
@@ -36,34 +44,30 @@
               (lam
                 n-7
                 (force
-                  (case
-                    (constr
-                      0
-                      (case
-                        (constr
-                          0
+                  (force
+                    [
+                      [
+                        [
+                          (force (builtin ifThenElse))
                           [ [ (builtin lessThanEqualsInteger) i-6 ] n-7 ]
-                          (con bool False)
-                          (con bool True)
-                        )
-                        (force (builtin ifThenElse))
-                      )
-                      (delay curr-5)
-                      (delay
-                        (case
-                          (constr
-                            0
-                            s-3
-                            curr-5
-                            [ [ (builtin addInteger) prev-4 ] curr-5 ]
-                            [ [ (builtin addInteger) (con integer 1) ] i-6 ]
-                            n-7
+                        ]
+                        (delay
+                          (delay
+                            [
+                              [
+                                [
+                                  [ [ s-3 s-3 ] curr-5 ]
+                                  [ [ (builtin addInteger) prev-4 ] curr-5 ]
+                                ]
+                                [ [ (builtin addInteger) (con integer 1) ] i-6 ]
+                              ]
+                              n-7
+                            ]
                           )
-                          s-3
                         )
-                      )
-                    )
-                    (force (builtin ifThenElse))
+                      ]
+                      (delay (delay curr-5))
+                    ]
                   )
                 )
               )
